Second try of wildcard support for testresult files (#435)
* Added semicolon and wildcard support for all call types (MsBuild, CommandLine, PowerShell, UI-api).

Folders and no match will deliver no result file.
Multiple matches on the same file will distinct into a single result.

* Improved design of tests for TestResultFiles in order to reflect their intention

* Changed add file into add directory to the filesystem mock.

Pickles throws an exception when a folder does not exist for a resultfile. The about selecting a non existing file was a victim of this exception making it fail. By ensuring the directory exists, the testcase passes.